requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
Django==1.11.6
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
httpie==0.9.9
idna==2.6
Pygments==2.2.0
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
urllib3==1.22
